Added a longer test for disc
The 10-step test in disc_short_10.dat is very short.  To properly test
a full timestep, the number of time steps needs to be at least
2^(number of levels) - in this case 1024, so a longer test with 2048
steps was made.

This test fails between timesteps 160 and 170 with a segfault when
using MPI+OpenMP but succeeds at least to timestep 290 when using
OpenMP.  It looks like some indices are not set correctly.  Note that
for MPI_OpenMP dt : 9.63153e-05 and for OpenMP dt : 0.009615.  The
changes made have been only to build and run Gandalf on DiAL, so it
seems unlikely that these will have caused this MPI error.