Use a register to hold the Box in ComputeGatherNeighbourList
Noticeable in the call of BoxOverlap from the loop over cells in
ComputeGatherNeighbourList (Tree/Tree.cpp) is that the in-lined code
uses only ymm0, even though the first box in the test is the same for
all iterations of the loop.  This means that there is a load from L1
each iteration.  Various compiler flags do not force the use of
registers, but the Gnu and Intel extension (for example)

  register __m256d gmin asm ("ymm14");

does (even though register is an unused and reserved keyword since
C++17).  Sort of:  the ymm14 and ymm15 registers were chosen so as not
to conflict with the xmm (128-bit) registers that are used, and that
alias to the lower half of the ymm (256-bit) registers; the in-lined
(here actually written in ComputeGatherNeighbourList) code still uses
ymm0, but does a move from the other ymm register.  The ymm14/15
registers need to be reloaded after the neibmanager.AddNeibs call -
that does a vzeroupper.

Using registers gives no speed-up.