Try prefetching the next cell's box
In ComputeGatherNeighbourList (Tree/Tree.cpp), the next cell's Box can
be usefully prefetched if the subtree is not too large (i.e. if the
prefetched Box will remain in cache).  This will only improve the
speed for L2 to L1, since the depth-first search of the kD-tree
already should be keeping most of the working set in L2 cache (L3 is
not inclusive on Skylake).

Testing this with the leaf cells (no subtree) gave a 20% speed-up of
ComputeGatherNeighbourList, so it may be useful to choose a higher
level in the tree, and apply this in the other Neighbour list
searches.  However, the speed-up seems to be partly from an
improvement in L2/L3 cache use.

Of course this requires visibility of the Box min and max arrays to
pass to _mm_prefetch (Box is standard layout, so the address of Box
could be used).