Always prefecth the next cell's data
Although there possibly is an optimum level of the tree below which
prefetching should be done, test by always prefetching.

The results are mixed.  The BoxOverlap times are halved but the time
is put back in the routines that call these:
ComputeGatherNeighbourList and ComputeNeighbourAndGhostList.  Possibly
the tests for cc < Ncell, or the calculation of the addresses to
prefetch, or the prefetch instructions themselves.