Test various prefetches for BoxOverlap
BoxOverlap is called in ComputeGatherNeighbourList,
ComputeNeighbourList, and ComputeNeighbourAndGhostList.  Explicit
prefetching does not help much - the time for BoxOveralps decrease (by
50%) but the time in the calling routine is increased by the same
amount.

Various tests were done:

T0 has the same effect as T1 (T1 is probably preferable when OpenMP is
used (untested), since it will not affect L3 cache).

The test for index < Ncell is probably not needed, the prefectch
instruction will not cause a fault if the address is invalid, but the
test has no effect on the times.

Doing everything except the prefetch itself (i.e. test for index <
Ncell, calculate the address of celldata[cc].cnext) makes the time in
the calling routine low, time in BoxOverlap middling, and time in
PeriodicBoxOverlap high (compared with the original code: low, high,
low).  The overal time is similar to the orogonal code and the
prefetched code.  It is unclear why there is extra time in
PeriodicBoxOverlap:  there are no periodic BCs.

Further tests using const, localising in a block, prefetching the
first child as well, not using the calcaulated next cell later in the
loop, all give no difference in the times.

Possible reasons for the lask of effect are:  1. the prefetch is too
late; 2. the prefetch interferes with hardware prefetch.