Fix strides
The Trres are constructed so that the array order mostly matches the
traversal in Compute...NeighbourList, i.e., celldata[c].copen will be
c+1, and for a leaf cell celldata[c].cnext will be c+1.  So choosing a
fixed stride whould help the prefetching.  In fact it has little
effect, and should have already been working well for
ComputeGatherNeighbourList for non-leaf cells, where the stride would
be one TreeCell.

However, looking at the addresses, the supposedly cacheline-aligned
objects are not aligned at all.  This was tested earlier for a simple
program but not checked in Gandalf itself.  The new for celldata (and
array of TreeCells) does not need to align the objects to extended
alignment, it aligns them to 16 bytes (C++ standard).  This will
invalidate all the cache_group_together tests.