Use software prefetch
Software prefetch a fixed number of elements of the Tree ahead.  This
avoids the read of the current TreeCell and indirection.  Reading 4
elements ahead gives about a 3% overall speed-up (about 30% speed-up
in BoxOverlap itself).  1,2,3,5 elements ahead do not give as much
speed-up.

Conditionals were re-ordered in BoxOverlap but this has no effect.