Add a check so that prefetch stays within the particle array
Keep the prefetches in _EndSearch within the particle array bounds.
This isn't actually necessary since _mm_preftech will not fail if the
address if not valid (but could be slow for address=0?).

Prefetches do not help in _EndSearch, tested for AHEAD_E=1,2,3,10.
Not enough non-memory work?  puahsbacks and writes to neibdata?