Print number of cells traversed, sizes of particle arrays
Print the number of cells traversed, the leaf cells to be used, the
number of particles to be searched, the sizes of the neibdata and
neiblist arrays.

These all look much larger than expected (i.e., thousands rather than
hundreds), which means that they will not fit in L2 cache and may not
fit in L2+L3.  That rather vitiates the attempts at improving L1 cache
performance, which has assumed that the working set will fit into L2.
It may be that the neibdata array will have to be bypassed, which
would slow down ComputeH and ComputeHydroForces - this would need
measured.