Use emplace_back for the list of neighbouring particles
Using emplace_back instead of push_back when adding DensityParticles
or HydroForcesParticles to neibdata shifts around counts in the VTune
analysis, but makes no difference to the overall time.

Note that counts (and times) can vary a lot as the code changes, but
this appears to be VTune ascribing the times differently.