Calculate frequency distributions of Nactive and Nculled
Calculate the frequency distribution of number of used SIMD lanes (if
vectorisation was used) for a/ the number of active particles in a
cell, and b/ the particles in the culled neighbour list.  The
assumption for the culled list is that the unculled list will be
processed sequentially by SIMD width, and the computations masked to
the particles in the culled; an alternative is to use the culled list
and Intel's gather operation.

a/ will give cache blocking, b/ will not, but the prefetcher on
Skylake is good (see previous tests, where software prefetching had no
effect), so it is likely that this will give no advantage.

Vectorisation in both cases will mean that the culled lists are not
used:  the gain from vectorisation may be less than the loss by having
to calculate for all the neighbour list (with masking).  The effect
will be strongest for the h computation, whre there is iteration, and
worse in this case for a/, where h for all the active particles will
be calculated until finally h has converged for all the particels,
even if several have already converged.  There will be a saving in
that the culled lists do not have to be made.

In case a/, the optimum Nleafmax for vectorisation may not be the
optimum Nleafmax for the tree.

Case a/ looks simpler to implement.

Tested for SIMD lengths of 4 and 8, for three cases

i. Nleafmax=8 and 8*2^17, which give an initial distribution with 8
particles in each cell.  For SIMD width 4, this estimates a speed up
of 4x for a/ and 3x for b/.  For SIMD width 8, this estimates a speed up
of 8x for a/ and 5x for b/.

ii. Nleafmax=8 and 9*2^17, which give an initial distribution with 4
or 5 particles in each cell.  For SIMD width 4, this estimates a speed up
of ?x for a/ and ?x for b/.  For SIMD width 8, this estimates a speed up
of 4x for a/ and 5x for b/.

iii. Nleafmax=6 and 1000000, which give an initial distribution with 3
or 4 particles in each cell.  For SIMD width 4, this estimates a speed up
of ?x for a/ and ?x for b/.  For SIMD width 8, this estimates a speed up
of 4x for a/ and 5x for b/.

Tests were for the disc case (GI_disc fails with no error message) and
for 128 or 2048 timesteps.

So, choice of Nleafmax and initial number of particles is important
(number of particles will change as particles accrete).