Add first steps of SIMD
SIMD parallelism over the active particles in a cell is being added.
This has probably better vectorisation than SIMD parallelism over the
neighbours, and also will give some cache blocking, but at the cost of
possible wasted iterations in ComputeH.  And it looks like some
operations have to be repeated in multiple neighbour loops, rather
than in the single TrimNeighbourLists function.

The first steps are to have arrays [0:n_active_particles-1] to
eventually be vectorised.

Start with ComputeH, then the hydrodynamical forces and then the
gravitational forces.

This commit has the array version of ComputeH commented out, to
compile, and no changes to the forces functions.