Array version of SIMD for ComputeH
An array version of SIMD for ComputeH is used.

The SIMD vector consists of the active particles in a cell.  This
means that the lopps are re-ordered from

create neighbour list (EndSearchGather)
loop over active particles
  loop over neighbour list to get reduced neighbour list (GetParticleNeibGather)
  loop over reduced neighbour list (possibly several times) to compute h

to

create neighbour list (EndSearchGather)
loop over neighbour list (possibly several times)
  loop over active particles

Since all the active particles are processed at once, the loop control
is more complex, using SIMD-wide mask arrays (eventually to be
replaced with bitmasks).  And the SIMD vectorisation is not yet
implemented (unknown if the compiler has auto-vectorised).

Currently, the test for a neighbour (in MaskParticleNeibGather), is
the same as in GetParticleNeibGather.

The neighbour list is 3x (for disc.dat) or 10x (for GI_disc.dat) as
long as the reduced neighbour list, so any gains from the cache
blocking (all active particles are done per neighbour) and
vectorisation may not be enough.  In addition, computing h is
iterative, and the loop will not complete until all active particles
are converged.

Tested with 10 time-steps of disc.dat, the time for SIMD ComputeH and
MaskParticleNeibGather is more then 2x the time for non-SIMD ComputeH
and GetParticleNeibGather.  Whether this can be improved remains to be
seen.

(GI_disc.dat no longer works since the correction to ComputeH for
particles in sinks.)